nova refatoração e implementação de listas

+++ src/com/screenmatch/calculations/Rateable.java
@@ -1,5 +1,5 @@
 package com.screenmatch.calculations;
 //interface avaliável, tornando um título possível de avaliação
 public interface Rateable {
-    int getRating();
+    double getRating();
 }
